Extract AssetSection to other file
